% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-package.R
\docType{package}
\name{lang-package}
\alias{lang}
\alias{lang-package}
\title{lang: Translates R Help Documentation using Large Language Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Translates R help documentation on the fly by using a Large Language model of your choice. If you are using 'RStudio' or 'Positron' the translated help will appear in the help pane.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlverse.github.io/lang/}
  \item \url{https://github.com/mlverse/lang}
  \item Report bugs at \url{https://github.com/mlverse/lang/issues}
}

}
\author{
\strong{Maintainer}: Edgar Ruiz \email{edgar@posit.co}

}
\keyword{internal}
