% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-help.R
\name{lang_help}
\alias{lang_help}
\title{Translates help}
\usage{
lang_help(
  topic,
  package = NULL,
  lang = Sys.getenv("LANG"),
  type = getOption("help_type")
)
}
\arguments{
\item{topic}{topic}

\item{package}{package}

\item{lang}{lang}

\item{type}{type}
}
\description{
Translates help
}
