% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-roxygen.R
\name{process_roxygen}
\alias{process_roxygen}
\title{Creates the Rd files based on translated Roxygen scripts}
\usage{
process_roxygen(lang, source = "man-lang", target = "inst/man-lang")
}
\arguments{
\item{lang}{2-letter language/source folder where the translated Roxygen
scripts are located}

\item{source}{Root source folder where the different translations are located.
Defaults to 'man-lang'.}

\item{target}{Root target folder where the different translations will be
located. Defaults to 'inst/man-lang'}
}
\value{
Multiple Rd files based on the source R scripts
}
\description{
Creates the Rd files based on translated Roxygen scripts
}
