% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-roxygen.R
\name{process_roxygen_folder}
\alias{process_roxygen_folder}
\alias{process_roxygen}
\title{Creates the Rd files based on translated Roxygen scripts}
\usage{
process_roxygen_folder(folder, source = "man-lang", target = "inst/man-lang")

process_roxygen(source = "man-lang", target = "inst/man-lang")
}
\arguments{
\item{folder}{Source sub-folder where the source Roxygen R scripts are}

\item{source}{Base source folder where the different translations are located.
Defaults to 'man-lang'.}

\item{target}{Base target folder where the different translations will be
located. Defaults to 'inst/man-lang'}
}
\value{
Multiple Rd files based on the source R scripts
}
\description{
Creates the Rd files based on translated Roxygen scripts
}
