% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso-639.R
\name{to_iso639}
\alias{to_iso639}
\alias{from_iso639}
\title{Convert to and from ISO 639 language code}
\usage{
to_iso639(lang, silent = TRUE)

from_iso639(iso)
}
\arguments{
\item{lang}{Name of the language to be converted}

\item{silent}{Flag to indicate if the function should return a message
with the result. If there is no match, it will be an error message}

\item{iso}{The two-letter ISO 639 code to search}
}
\description{
Convert to and from ISO 639 language code
}
