% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-pkgload.R
\name{help}
\alias{help}
\alias{shim_help}
\alias{?}
\alias{shim_question}
\title{Drop-in replacements for help and ? functions}
\usage{
# help(topic, package = NULL, ...)

# ?e2
# e1?e2
}
\arguments{
\item{topic}{A name or character string specifying the help topic.}

\item{package}{A name or character string specifying the package in which
to search for the help topic. If NULL, search all packages.}

\item{...}{Additional arguments to pass to \code{\link[utils:help]{utils::help()}}.}

\item{e1}{First argument to pass along to \verb{utils::}?``.}

\item{e2}{Second argument to pass along to \verb{utils::}?``.}
}
\description{
The \verb{?} and \code{help} functions are replacements for functions of the
same name in the utils package. If the LANG environment variable is not set
to English, it will activate the translation to whatever language LANG is
set to.
}
