% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-help.R
\name{lang_help}
\alias{lang_help}
\title{Translates help}
\usage{
lang_help(topic, package = NULL, lang = NULL, type = getOption("help_type"))
}
\arguments{
\item{topic}{The topic to search for}

\item{package}{The R package to look for the topic}

\item{lang}{Language to translate the help to}

\item{type}{Produce "html" or "text" output for the help. It default to
\code{getOption("help_type")}}
}
\description{
Translates help
}
\examples{
\donttest{
library(lang)
llm_use("ollama", "llama3.2", .silent = TRUE)
lang_help("aes", "ggplot2", lang = "french")
}

}
