% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-help.R
\name{lang_help}
\alias{lang_help}
\title{Translates help}
\usage{
lang_help(
  topic,
  package = NULL,
  lang = NULL,
  type = getOption("help_type"),
  package_rds = "man-lang"
)
}
\arguments{
\item{topic}{The topic to search for}

\item{package}{The R package to look for the topic}

\item{lang}{Language to translate the help to}

\item{type}{Produce "html" or "text" output for the help. It default to
\code{getOption("help_type")}}
}
\value{
Returns either a text or html version of the topic's help
documentation
}
\description{
Translates a given topic into a target language. It uses the \code{lang} argument
to determine which language to translate to. If not passed, this function will
look for a target language in the LANG and LANGUAGE environment variables to
determine the target language. If the target language is English, no translation
will be processed, so the help returned will be the original package's
documentation.
}
\examples{
\donttest{
library(lang)
llm_use("ollama", "llama3.2", .silent = TRUE)
lang_help("aes", "ggplot2", lang = "french")
}
}
