% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-roxygen.R
\name{process_roxygen}
\alias{process_roxygen}
\title{Creates the Rd files based on translated Roxygen scripts}
\usage{
process_roxygen(
  source_sub_folder = NULL,
  source_folder = "man-lang",
  target_folder = "inst/man-lang",
  r_folder = "R",
  pkg_path = "."
)
}
\arguments{
\item{source_sub_folder}{Source sub-folder where the source Roxygen R scripts
are. Defaults to NULL. If left null, all of the sub-folders in the
\code{source_folder} will be processed}

\item{source_folder}{Base source folder where the different translations are located.
Defaults to 'man-lang'.}

\item{target_folder}{Base target folder where the different translations will be
located. Defaults to 'inst/man-lang'}

\item{r_folder}{Source of the original R scripts. Only used to see if the
Roxygen documentation is different from what is capture in the \code{source_folder}}

\item{pkg_path}{The path to the package}
}
\value{
Multiple Rd files based on the source R scripts
}
\description{
Creates the Rd files based on translated Roxygen scripts
}
