% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two-letter.R
\name{to_iso639}
\alias{to_iso639}
\alias{from_iso639}
\title{Convert to and from ISO 639 language code}
\usage{
to_iso639(lang)

from_iso639(iso)
}
\arguments{
\item{lang}{Name of the language to be converted}

\item{iso}{The two-letter ISO 639 code to search}
}
\description{
Convert to and from ISO 639 language code
}
