% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-roxygen.R
\name{translate_roxygen}
\alias{translate_roxygen}
\title{Translates the Roxygen2 documentation}
\usage{
translate_roxygen(
  lang,
  target_subfolder,
  target = path("man-lang"),
  source = path("R")
)
}
\arguments{
\item{lang}{The target language to translate help to}

\item{target_subfolder}{2-letter language/source folder to save the new
Roxygen scripts to}

\item{target}{The target base folder to save the Roxygen files. It defaults
to 'man-lang'. The final destination will be a combination of this and the
folder from \code{target_subfolder}}

\item{source}{The source R scripts. This can be a folder or a single file. It
defaults to the 'R' folder.}
}
\description{
Reads the Roxygen2 tags in the package and translates them. The translations
are stored in R scripts. The default location of the new scripts is 'man-lang'.
The will be in a sub-folder representing the language the are translated to.
}
\details{
This approach makes it easier to edit the translations by hand after
the LLM does a first pass. This way it is easier for others to collaborate
with improving the translation
}
