% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-roxygen.R
\name{translate_roxygen}
\alias{translate_roxygen}
\title{Translates the Roxygen2 documentation}
\usage{
translate_roxygen(lang = NULL, path = "R", dir = fs::path("man-lang", lang))
}
\arguments{
\item{lang}{2-letter target language to translate to}

\item{path}{The source R scripts. This can be a folder or a single file. It
defaults to the R folder.}

\item{dir}{The target folder to save the new, translated, R scripts to. It
defaults to 'man-lan/\link{2 letter target language}'}
}
\description{
Reads the Roxygen2 tags in the package and translates them. The translations
are stored in R scripts. The default location of the new scripts is 'man-lang'.
The will be in a sub-folder representing the language the are translated to.
}
\details{
This approach makes it easier to edit the translations by hand after
the LLM does a first pass. This way it is easier for others to collaborate
with improving the translation
}
